+++ src/components/Breadcrumb/index.tsx
@@ -10,6 +10,7 @@ import { BreadcrumbSchema, Enum } from "../../types";
 import { CONTAINER_SIZE } from "../../constants";
 import { Badge, BadgeProps } from "../Badge";
 import "./index.scss";
+import "../../styles/icons/fa_icons.css";
 
 export interface BreadcrumbProps {
   breadcrumbLists: BreadcrumbSchema[];