+++ src/components/Dropdown/DropdownItemGroup.stories.tsx
@@ -6,6 +6,7 @@
 
 import { Meta, Story } from "@storybook/react/types-6-0";
 import { DropdownItemGroup, DropdownItem, DropdownItemGroupProps, ButtonDropdown } from ".";
+import "../../styles/icons/fa_icons.css";
 
 export default {
   title: "Components/Dropdown/DropdownItemGroup",