+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Refactored
+- Import fa_icons.css in all components using `fa-icon`
+
 ## [0.6.5]
 ### Refactored
 - Remove `styled-atlaskit`, `Form`, `Loading`, `Navigation`, `Page`, and `Wrapper` folder and move its components up by one level